move optimize serialization tests to kernel tests
